Calculate time for ebb and flow at a other place
The ebb and flow for different standard harbours are given. The time for
ebb and flow for other harbours can be calculated based on minutes to
add or subtract, a height correction factor, and a fixed delta height to
add to the calculated height.

Timepoints in ISO format and height in cm are written for each timepoint
on a line in a file with the standard harbour as name
(kristandsundiso.txt). The program output the ebb and flow for the
secondary harbour in a new file named after the harbour (uthaugiso.txt).

Standard and secondary harbours, minutes to adjust, height correction
factor and the fixed delta to add, are all written in a configuration
file ebb.cfg.